find_venv_bin function to common.py
