+++ pyempaq/common.py
@@ -1,0 +1,20 @@
+# Copyright 2021 Facundo Batista
+# Licensed under the GPL v3 License
+# For further info, check https://github.com/facundobatista/pyempaq
+
+"""Common functionality."""
+
+def find_venv_bin(basedir, exec_base):
+    """Heuristics to find the pip executable in different platforms."""
+    bin_dir = basedir / "bin"
+    if bin_dir.exists():
+        # linux-like environment
+        return bin_dir / exec_base
+
+    bin_dir = basedir / "Scripts"
+    if bin_dir.exists():
+        # windows environment
+        return bin_dir / f"{exec_base}.exe"
+
+    raise RuntimeError(f"Binary not found inside venv; subdirs: {list(basedir.iterdir())}")
+
